Initial pass at GVParse
Generators have now been implemenented.
DC Lines implemenented

Note that Ancillary services are NOT YET covered, but there are some
notes in the README about where the data is.

Note: only Battery model implemented, Pumped storage to come.

under data_model_tests is a test script called test_gvparser.py that
takes the path to an h5 file as input.
If this is pointed to the h5 for the WECC 240 model it should complete
successfully.