Adding distributed generation to load
distributed generation to load comes in two types:
1. AREA type in GenerationDistribution or
2. in BTMGenLoadMapping

The add_load method now calls to functions to go through these tables
and modify the load.
a list is also kept of all generators that are modifying any given load
mainly for information/debugging purposes.

Note: this could alternatively be modeled via a distributed generator.
One possibility is to look at generator type, if it is hourly, do it
this way, otherwise use the distributed generation capability from
Brent.
That should probably be a configuration option at that point.