Working conversion to AC model
The GVParser can now handle multiple exceptions to be able to generate
some sort of an output for the WECC cases.
This relies on essentially fixing results to the output when a model
is not yet implemented.

Distributed generators are now handled.

utility is written to "flatten" distributed generators.
This is useful if for some reason we don't want to model them together,
or, when trying to match with power flow model where the concept of
distributed generation doesn't exist.

The power world parser takes an egret model and addes reactive power
elements and limits.
Again, error handling allows to simply skip any missed elements.

IMPORTANT: the objective of the ACOPF solution is NOT to generate a
perfectly working case but rather to generate a solution that should be
close to a good one. This way, when the set points are moved to the
Power Flow solver, hopefully newton-raphson simply converges.